+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/engine/EngineObserver.kt
@@ -111,7 +111,7 @@ internal class EngineObserver(
 
     override fun onExternalResource(
         url: String,
-        fileName: String,
+        fileName: String?,
         contentLength: Long?,
         contentType: String?,
         cookie: String?,