+++ mobile/android/android-components/components/concept/fetch/src/main/java/mozilla/components/concept/fetch/Headers.kt
@@ -159,3 +159,5 @@ class MutableHeaders(headers: List<Header>) : Headers, MutableIterable<Header> {
 
     override fun hashCode() = headers.hashCode()
 }
+
+fun List<Header>.toMutableHeaders() = MutableHeaders(this)