+++ mobile/android/android-components/components/support/utils/src/main/java/mozilla/components/support/utils/DownloadUtils.kt
@@ -61,35 +61,21 @@ object DownloadUtils {
     @JvmStatic
     @SuppressWarnings("ComplexMethod", "LongMethod", "NestedBlockDepth")
     fun guessFileName(contentDisposition: String?, url: String?, mimeType: String?): String {
+        val mimeTypeMap = MimeTypeMap.getSingleton()
         var filename: String? = null
         var extension: String? = null
 
         // Extract file name from content disposition header field
         if (contentDisposition != null) {
-            filename = parseContentDisposition(contentDisposition)
-            if (filename != null) {
-                val index = filename.lastIndexOf('/') + 1
-                if (index > 0) {
-                    filename = filename.substring(index)
-                }
-            }
+            filename = parseContentDisposition(contentDisposition)?.substringAfterLast('/')
         }
 
         // If all the other http-related approaches failed, use the plain uri
         if (filename == null) {
-            var decodedUrl: String? = Uri.decode(url)
-            if (decodedUrl != null) {
-                val queryIndex = decodedUrl.indexOf('?')
-                // If there is a query string strip it, same as desktop browsers
-                if (queryIndex > 0) {
-                    decodedUrl = decodedUrl.substring(0, queryIndex)
-                }
-                if (!decodedUrl.endsWith("/")) {
-                    val index = decodedUrl.lastIndexOf('/') + 1
-                    if (index > 0) {
-                        filename = decodedUrl.substring(index)
-                    }
-                }
+            // If there is a query string strip it, same as desktop browsers
+            val decodedUrl: String? = Uri.decode(url)?.substringBefore('?')
+            if (decodedUrl?.endsWith('/') == false) {
+                filename = decodedUrl.substringAfterLast('/')
             }
         }
 
@@ -103,10 +89,7 @@ object DownloadUtils {
         val dotIndex = filename.indexOf('.')
         if (dotIndex < 0) {
             if (mimeType != null) {
-                extension = MimeTypeMap.getSingleton().getExtensionFromMimeType(mimeType)
-                if (extension != null) {
-                    extension = "" + extension
-                }
+                extension = mimeTypeMap.getExtensionFromMimeType(mimeType)?.let { ".$it" }
             }
             if (extension == null) {
                 extension = if (mimeType?.toLowerCase(Locale.ROOT)?.startsWith("text/") == true) {
@@ -123,14 +106,9 @@ object DownloadUtils {
             if (mimeType != null) {
                 // Compare the last segment of the extension against the mime type.
                 // If there's a mismatch, discard the entire extension.
-                val lastDotIndex = filename.lastIndexOf('.')
-                val typeFromExt = MimeTypeMap.getSingleton().getMimeTypeFromExtension(
-                        filename.substring(lastDotIndex + 1))
-                if (typeFromExt == null || !typeFromExt.equals(mimeType, ignoreCase = true)) {
-                    extension = MimeTypeMap.getSingleton().getExtensionFromMimeType(mimeType)
-                    if (extension != null) {
-                        extension = ".$extension"
-                    }
+                val typeFromExt = mimeTypeMap.getMimeTypeFromExtension(filename.substringAfterLast('.'))
+                if (typeFromExt?.equals(mimeType, ignoreCase = true) != false) {
+                    extension = mimeTypeMap.getExtensionFromMimeType(mimeType)?.let { ".$it" }
                 }
             }
             if (extension == null) {